% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Apk}
\alias{Apk}
\title{Google Play Developer API Objects 
Lets Android application developers access their Google Play accounts.}
\usage{
Apk(binary = NULL, versionCode = NULL)
}
\arguments{
\item{binary}{Information about the binary payload of this APK}

\item{versionCode}{The version code of the APK, as specified in the APK's manifest file}
}
\value{
Apk object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:27:00
filename: /Users/mark/dev/R/autoGoogleAPI/googleandroidpublisherv2.auto/R/androidpublisher_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Apk Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
