% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ApkBinary}
\alias{ApkBinary}
\title{ApkBinary Object}
\usage{
ApkBinary(sha1 = NULL)
}
\arguments{
\item{sha1}{A sha1 hash of the APK payload, encoded as a hex string and matching the output of the sha1sum command}
}
\value{
ApkBinary object
}
\description{
ApkBinary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the binary payload of an APK.
}
