% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ApkListing}
\alias{ApkListing}
\title{ApkListing Object}
\usage{
ApkListing(language = NULL, recentChanges = NULL)
}
\arguments{
\item{language}{The language code, in BCP 47 format (eg 'en-US')}

\item{recentChanges}{Describe what's new in your APK}
}
\value{
ApkListing object
}
\description{
ApkListing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
