% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ApkListingsListResponse}
\alias{ApkListingsListResponse}
\title{ApkListingsListResponse Object}
\usage{
ApkListingsListResponse(listings = NULL)
}
\arguments{
\item{listings}{No description}
}
\value{
ApkListingsListResponse object
}
\description{
ApkListingsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

