% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ApksAddExternallyHostedRequest}
\alias{ApksAddExternallyHostedRequest}
\title{ApksAddExternallyHostedRequest Object}
\usage{
ApksAddExternallyHostedRequest(externallyHostedApk = NULL)
}
\arguments{
\item{externallyHostedApk}{The definition of the externally-hosted APK and where it is located}
}
\value{
ApksAddExternallyHostedRequest object
}
\description{
ApksAddExternallyHostedRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
