% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ApksAddExternallyHostedResponse}
\alias{ApksAddExternallyHostedResponse}
\title{ApksAddExternallyHostedResponse Object}
\usage{
ApksAddExternallyHostedResponse(externallyHostedApk = NULL)
}
\arguments{
\item{externallyHostedApk}{The definition of the externally-hosted APK and where it is located}
}
\value{
ApksAddExternallyHostedResponse object
}
\description{
ApksAddExternallyHostedResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

