% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{AppDetails}
\alias{AppDetails}
\title{AppDetails Object}
\usage{
AppDetails(contactEmail = NULL, contactPhone = NULL,
  contactWebsite = NULL, defaultLanguage = NULL)
}
\arguments{
\item{contactEmail}{The user-visible support email for this app}

\item{contactPhone}{The user-visible support telephone number for this app}

\item{contactWebsite}{The user-visible website for this app}

\item{defaultLanguage}{Default language code, in BCP 47 format (eg 'en-US')}
}
\value{
AppDetails object
}
\description{
AppDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
