% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{AppEdit}
\alias{AppEdit}
\title{AppEdit Object}
\usage{
AppEdit(expiryTimeSeconds = NULL, id = NULL)
}
\arguments{
\item{expiryTimeSeconds}{The time at which the edit will expire and will be no longer valid for use in any subsequent API calls (encoded as seconds since the Epoch)}

\item{id}{The ID of the edit that can be used in subsequent API calls}
}
\value{
AppEdit object
}
\description{
AppEdit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an edit of an app. An edit allows clients to make multiple changes before committing them in one operation.
}
\seealso{
Other AppEdit functions: \code{\link{edits.insert}}
}
