% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Comment}
\alias{Comment}
\title{Comment Object}
\usage{
Comment(developerComment = NULL, userComment = NULL)
}
\arguments{
\item{developerComment}{A comment from a developer}

\item{userComment}{A comment from a user}
}
\value{
Comment object
}
\description{
Comment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
