% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{DeobfuscationFile}
\alias{DeobfuscationFile}
\title{DeobfuscationFile Object}
\usage{
DeobfuscationFile(symbolType = NULL)
}
\arguments{
\item{symbolType}{The type of the deobfuscation file}
}
\value{
DeobfuscationFile object
}
\description{
DeobfuscationFile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a deobfuscation file.
}
