% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{DeveloperComment}
\alias{DeveloperComment}
\title{DeveloperComment Object}
\usage{
DeveloperComment(lastModified = NULL, text = NULL)
}
\arguments{
\item{lastModified}{The last time at which this comment was updated}

\item{text}{The content of the comment, i}
}
\value{
DeveloperComment object
}
\description{
DeveloperComment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
