% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{DeviceMetadata}
\alias{DeviceMetadata}
\title{DeviceMetadata Object}
\usage{
DeviceMetadata(cpuMake = NULL, cpuModel = NULL, deviceClass = NULL,
  glEsVersion = NULL, manufacturer = NULL, nativePlatform = NULL,
  productName = NULL, ramMb = NULL, screenDensityDpi = NULL,
  screenHeightPx = NULL, screenWidthPx = NULL)
}
\arguments{
\item{cpuMake}{Device CPU make e}

\item{cpuModel}{Device CPU model e}

\item{deviceClass}{Device class (e}

\item{glEsVersion}{OpenGL version}

\item{manufacturer}{Device manufacturer (e}

\item{nativePlatform}{Comma separated list of native platforms (e}

\item{productName}{Device model name (e}

\item{ramMb}{Device RAM in Megabytes e}

\item{screenDensityDpi}{Screen density in DPI}

\item{screenHeightPx}{Screen height in pixels}

\item{screenWidthPx}{Screen width in pixels}
}
\value{
DeviceMetadata object
}
\description{
DeviceMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
