% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_functions.R
\name{edits.insert}
\alias{edits.insert}
\title{Creates a new edit for an app, populated with the app's current state.}
\usage{
edits.insert(AppEdit, packageName)
}
\arguments{
\item{AppEdit}{The \link{AppEdit} object to pass to this method}

\item{packageName}{Unique identifier for the Android app that is being updated; for example, 'com}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android-publisher}{Google Documentation}

Other AppEdit functions: \code{\link{AppEdit}}
}
