% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_functions.R
\name{edits.validate}
\alias{edits.validate}
\title{Checks that the edit can be successfully committed. The edit's changes are not applied to the live app.}
\usage{
edits.validate(packageName, editId)
}
\arguments{
\item{packageName}{Unique identifier for the Android app that is being updated; for example, 'com}

\item{editId}{Unique identifier for this edit}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android-publisher}{Google Documentation}
}

