% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Entitlement}
\alias{Entitlement}
\title{Entitlement Object}
\usage{
Entitlement(productId = NULL, productType = NULL, token = NULL)
}
\arguments{
\item{productId}{The SKU of the product}

\item{productType}{The type of the inapp product}

\item{token}{The token which can be verified using the subscriptions or products API}
}
\value{
Entitlement object
}
\description{
Entitlement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Entitlement resource indicates a user's current entitlement to an inapp item or subscription.
}
