% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_functions.R
\name{entitlements.list}
\alias{entitlements.list}
\title{Lists the user's current inapp item or subscription entitlements}
\usage{
entitlements.list(packageName, maxResults = NULL, productId = NULL,
  startIndex = NULL, token = NULL)
}
\arguments{
\item{packageName}{The package name of the application the inapp product was sold in (for example, 'com}

\item{maxResults}{}

\item{productId}{The product id of the inapp product (for example, 'sku1')}

\item{startIndex}{}

\item{token}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android-publisher}{Google Documentation}
}

