% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{EntitlementsListResponse}
\alias{EntitlementsListResponse}
\title{EntitlementsListResponse Object}
\usage{
EntitlementsListResponse(pageInfo = NULL, resources = NULL,
  tokenPagination = NULL)
}
\arguments{
\item{pageInfo}{No description}

\item{resources}{No description}

\item{tokenPagination}{No description}
}
\value{
EntitlementsListResponse object
}
\description{
EntitlementsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
