% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ExpansionFile}
\alias{ExpansionFile}
\title{ExpansionFile Object}
\usage{
ExpansionFile(fileSize = NULL, referencesVersion = NULL)
}
\arguments{
\item{fileSize}{If set this field indicates that this APK has an Expansion File uploaded to it: this APK does not reference another APK's Expansion File}

\item{referencesVersion}{If set this APK's Expansion File references another APK's Expansion File}
}
\value{
ExpansionFile object
}
\description{
ExpansionFile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
