% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ExternallyHostedApk}
\alias{ExternallyHostedApk}
\title{ExternallyHostedApk Object}
\usage{
ExternallyHostedApk(applicationLabel = NULL, certificateBase64s = NULL,
  externallyHostedUrl = NULL, fileSha1Base64 = NULL,
  fileSha256Base64 = NULL, fileSize = NULL, iconBase64 = NULL,
  maximumSdk = NULL, minimumSdk = NULL, nativeCodes = NULL,
  packageName = NULL, usesFeatures = NULL, usesPermissions = NULL,
  versionCode = NULL, versionName = NULL)
}
\arguments{
\item{applicationLabel}{The application label}

\item{certificateBase64s}{A certificate (or array of certificates if a certificate-chain is used) used to signed this APK, represented as a base64 encoded byte array}

\item{externallyHostedUrl}{The URL at which the APK is hosted}

\item{fileSha1Base64}{The SHA1 checksum of this APK, represented as a base64 encoded byte array}

\item{fileSha256Base64}{The SHA256 checksum of this APK, represented as a base64 encoded byte array}

\item{fileSize}{The file size in bytes of this APK}

\item{iconBase64}{The icon image from the APK, as a base64 encoded byte array}

\item{maximumSdk}{The maximum SDK supported by this APK (optional)}

\item{minimumSdk}{The minimum SDK targeted by this APK}

\item{nativeCodes}{The native code environments supported by this APK (optional)}

\item{packageName}{The package name}

\item{usesFeatures}{The features required by this APK (optional)}

\item{usesPermissions}{The permissions requested by this APK}

\item{versionCode}{The version code of this APK}

\item{versionName}{The version name of this APK}
}
\value{
ExternallyHostedApk object
}
\description{
ExternallyHostedApk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines an APK available for this application that is hosted externally and not uploaded to Google Play. This function is only available to enterprises who are using Google Play for Work, and whos application is restricted to the enterprise private channel
}

