% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ExternallyHostedApkUsesPermission}
\alias{ExternallyHostedApkUsesPermission}
\title{ExternallyHostedApkUsesPermission Object}
\usage{
ExternallyHostedApkUsesPermission(maxSdkVersion = NULL, name = NULL)
}
\arguments{
\item{maxSdkVersion}{Optionally, the maximum SDK version for which the permission is required}

\item{name}{The name of the permission requested}
}
\value{
ExternallyHostedApkUsesPermission object
}
\description{
ExternallyHostedApkUsesPermission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A permission used by this APK.
}

