% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(id = NULL, sha1 = NULL, url = NULL)
}
\arguments{
\item{id}{A unique id representing this image}

\item{sha1}{A sha1 hash of the image that was uploaded}

\item{url}{A URL that will serve a preview of the image}
}
\value{
Image object
}
\description{
Image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

