% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{InAppProduct.listings}
\alias{InAppProduct.listings}
\title{InAppProduct.listings Object}
\usage{
InAppProduct.listings()
}
\value{
InAppProduct.listings object
}
\description{
InAppProduct.listings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List of localized title and description data.
}
\seealso{
Other InAppProduct functions: \code{\link{InAppProduct.prices}},
  \code{\link{InAppProduct}},
  \code{\link{inappproducts.insert}},
  \code{\link{inappproducts.patch}},
  \code{\link{inappproducts.update}}
}
