% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{InAppProduct.prices}
\alias{InAppProduct.prices}
\title{InAppProduct.prices Object}
\usage{
InAppProduct.prices()
}
\value{
InAppProduct.prices object
}
\description{
InAppProduct.prices Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prices per buyer region. None of these prices should be zero. In-app products can never be free.
}
\seealso{
Other InAppProduct functions: \code{\link{InAppProduct.listings}},
  \code{\link{InAppProduct}},
  \code{\link{inappproducts.insert}},
  \code{\link{inappproducts.patch}},
  \code{\link{inappproducts.update}}
}
