% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{InAppProduct}
\alias{InAppProduct}
\title{InAppProduct Object}
\usage{
InAppProduct(InAppProduct.listings = NULL, InAppProduct.prices = NULL,
  defaultLanguage = NULL, defaultPrice = NULL, listings = NULL,
  packageName = NULL, prices = NULL, purchaseType = NULL, season = NULL,
  sku = NULL, status = NULL, subscriptionPeriod = NULL,
  trialPeriod = NULL)
}
\arguments{
\item{InAppProduct.listings}{The \link{InAppProduct.listings} object or list of objects}

\item{InAppProduct.prices}{The \link{InAppProduct.prices} object or list of objects}

\item{defaultLanguage}{The default language of the localized data, as defined by BCP 47}

\item{defaultPrice}{Default price cannot be zero}

\item{listings}{List of localized title and description data}

\item{packageName}{The package name of the parent app}

\item{prices}{Prices per buyer region}

\item{purchaseType}{Purchase type enum value}

\item{season}{Definition of a season for a seasonal subscription}

\item{sku}{The stock-keeping-unit (SKU) of the product, unique within an app}

\item{status}{No description}

\item{subscriptionPeriod}{Subscription period, specified in ISO 8601 format}

\item{trialPeriod}{Trial period, specified in ISO 8601 format}
}
\value{
InAppProduct object
}
\description{
InAppProduct Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InAppProduct functions: \code{\link{InAppProduct.listings}},
  \code{\link{InAppProduct.prices}},
  \code{\link{inappproducts.insert}},
  \code{\link{inappproducts.patch}},
  \code{\link{inappproducts.update}}
}

