% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{InAppProductListing}
\alias{InAppProductListing}
\title{InAppProductListing Object}
\usage{
InAppProductListing(description = NULL, title = NULL)
}
\arguments{
\item{description}{No description}

\item{title}{No description}
}
\value{
InAppProductListing object
}
\description{
InAppProductListing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

