% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_functions.R
\name{inappproducts.delete}
\alias{inappproducts.delete}
\title{Delete an in-app product for an app.}
\usage{
inappproducts.delete(packageName, sku)
}
\arguments{
\item{packageName}{Unique identifier for the Android app with the in-app product; for example, 'com}

\item{sku}{Unique identifier for the in-app product}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android-publisher}{Google Documentation}
}
