% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Listing}
\alias{Listing}
\title{Listing Object}
\usage{
Listing(fullDescription = NULL, language = NULL, shortDescription = NULL,
  title = NULL, video = NULL)
}
\arguments{
\item{fullDescription}{Full description of the app; this may be up to 4000 characters in length}

\item{language}{Language localization code (for example, 'de-AT' for Austrian German)}

\item{shortDescription}{Short description of the app (previously known as promo text); this may be up to 80 characters in length}

\item{title}{App's localized title}

\item{video}{URL of a promotional YouTube video for the app}
}
\value{
Listing object
}
\description{
Listing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

