% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ListingsListResponse}
\alias{ListingsListResponse}
\title{ListingsListResponse Object}
\usage{
ListingsListResponse(listings = NULL)
}
\arguments{
\item{listings}{No description}
}
\value{
ListingsListResponse object
}
\description{
ListingsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

