% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{MonthDay}
\alias{MonthDay}
\title{MonthDay Object}
\usage{
MonthDay(day = NULL, month = NULL)
}
\arguments{
\item{day}{Day of a month, value in [1, 31] range}

\item{month}{Month of a year}
}
\value{
MonthDay object
}
\description{
MonthDay Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
