% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Price}
\alias{Price}
\title{Price Object}
\usage{
Price(currency = NULL, priceMicros = NULL)
}
\arguments{
\item{currency}{3 letter Currency code, as defined by ISO 4217}

\item{priceMicros}{The price in millionths of the currency base unit represented as a string}
}
\value{
Price object
}
\description{
Price Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
