% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ProductPurchase}
\alias{ProductPurchase}
\title{ProductPurchase Object}
\usage{
ProductPurchase(consumptionState = NULL, developerPayload = NULL,
  purchaseState = NULL, purchaseTimeMillis = NULL)
}
\arguments{
\item{consumptionState}{The consumption state of the inapp product}

\item{developerPayload}{A developer-specified string that contains supplemental information about an order}

\item{purchaseState}{The purchase state of the order}

\item{purchaseTimeMillis}{The time the product was purchased, in milliseconds since the epoch (Jan 1, 1970)}
}
\value{
ProductPurchase object
}
\description{
ProductPurchase Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A ProductPurchase resource indicates the status of a user's inapp product purchase.
}
