% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Prorate}
\alias{Prorate}
\title{Prorate Object}
\usage{
Prorate(defaultPrice = NULL, start = NULL)
}
\arguments{
\item{defaultPrice}{Default price cannot be zero and must be less than the full subscription price}

\item{start}{Defines the first day on which the price takes effect}
}
\value{
Prorate object
}
\description{
Prorate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

