% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_functions.R
\name{purchases.subscriptions.defer}
\alias{purchases.subscriptions.defer}
\title{Defers a user's subscription purchase until a specified future expiration time.}
\usage{
purchases.subscriptions.defer(SubscriptionPurchasesDeferRequest, packageName,
  subscriptionId, token)
}
\arguments{
\item{SubscriptionPurchasesDeferRequest}{The \link{SubscriptionPurchasesDeferRequest} object to pass to this method}

\item{packageName}{The package name of the application for which this subscription was purchased (for example, 'com}

\item{subscriptionId}{The purchased subscription ID (for example, 'monthly001')}

\item{token}{The token provided to the user's device when the subscription was purchased}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android-publisher}{Google Documentation}

Other SubscriptionPurchasesDeferRequest functions: \code{\link{SubscriptionPurchasesDeferRequest}}
}
