% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_functions.R
\name{purchases.voidedpurchases.list}
\alias{purchases.voidedpurchases.list}
\title{Lists the purchases that were cancelled, refunded or charged-back.}
\usage{
purchases.voidedpurchases.list(packageName, endTime = NULL,
  maxResults = NULL, startIndex = NULL, startTime = NULL, token = NULL)
}
\arguments{
\item{packageName}{The package name of the application for which voided purchases need to be returned (for example, 'com}

\item{endTime}{The time, in milliseconds since the Epoch, of the newest voided in-app product purchase that you want to see in the response}

\item{maxResults}{}

\item{startIndex}{}

\item{startTime}{The time, in milliseconds since the Epoch, of the oldest voided in-app product purchase that you want to see in the response}

\item{token}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android-publisher}{Google Documentation}
}
