% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Review}
\alias{Review}
\title{Review Object}
\usage{
Review(authorName = NULL, comments = NULL, reviewId = NULL)
}
\arguments{
\item{authorName}{The name of the user who wrote the review}

\item{comments}{A repeated field containing comments for the review}

\item{reviewId}{Unique identifier for this review}
}
\value{
Review object
}
\description{
Review Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

