% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ReviewReplyResult}
\alias{ReviewReplyResult}
\title{ReviewReplyResult Object}
\usage{
ReviewReplyResult(lastEdited = NULL, replyText = NULL)
}
\arguments{
\item{lastEdited}{The time at which the reply took effect}

\item{replyText}{The reply text that was applied}
}
\value{
ReviewReplyResult object
}
\description{
ReviewReplyResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

