% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_functions.R
\name{reviews.list}
\alias{reviews.list}
\title{Returns a list of reviews. Only reviews from last week will be returned.}
\usage{
reviews.list(packageName, maxResults = NULL, startIndex = NULL,
  token = NULL, translationLanguage = NULL)
}
\arguments{
\item{packageName}{Unique identifier for the Android app for which we want reviews; for example, 'com}

\item{maxResults}{}

\item{startIndex}{}

\item{token}{}

\item{translationLanguage}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android-publisher}{Google Documentation}
}
