% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ReviewsListResponse}
\alias{ReviewsListResponse}
\title{ReviewsListResponse Object}
\usage{
ReviewsListResponse(pageInfo = NULL, reviews = NULL,
  tokenPagination = NULL)
}
\arguments{
\item{pageInfo}{No description}

\item{reviews}{No description}

\item{tokenPagination}{No description}
}
\value{
ReviewsListResponse object
}
\description{
ReviewsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
