% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ReviewsReplyRequest}
\alias{ReviewsReplyRequest}
\title{ReviewsReplyRequest Object}
\usage{
ReviewsReplyRequest(replyText = NULL)
}
\arguments{
\item{replyText}{The text to set as the reply}
}
\value{
ReviewsReplyRequest object
}
\description{
ReviewsReplyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ReviewsReplyRequest functions: \code{\link{reviews.reply}}
}
