% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{ReviewsReplyResponse}
\alias{ReviewsReplyResponse}
\title{ReviewsReplyResponse Object}
\usage{
ReviewsReplyResponse(result = NULL)
}
\arguments{
\item{result}{No description}
}
\value{
ReviewsReplyResponse object
}
\description{
ReviewsReplyResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
