% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Season}
\alias{Season}
\title{Season Object}
\usage{
Season(end = NULL, prorations = NULL, start = NULL)
}
\arguments{
\item{end}{Inclusive end date of the recurrence period}

\item{prorations}{Optionally present list of prorations for the season}

\item{start}{Inclusive start date of the recurrence period}
}
\value{
Season object
}
\description{
Season Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
