% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{SubscriptionDeferralInfo}
\alias{SubscriptionDeferralInfo}
\title{SubscriptionDeferralInfo Object}
\usage{
SubscriptionDeferralInfo(desiredExpiryTimeMillis = NULL,
  expectedExpiryTimeMillis = NULL)
}
\arguments{
\item{desiredExpiryTimeMillis}{The desired next expiry time to assign to the subscription, in milliseconds since the Epoch}

\item{expectedExpiryTimeMillis}{The expected expiry time for the subscription}
}
\value{
SubscriptionDeferralInfo object
}
\description{
SubscriptionDeferralInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A SubscriptionDeferralInfo contains the data needed to defer a subscription purchase to a future expiry time.
}
