% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{SubscriptionPurchasesDeferRequest}
\alias{SubscriptionPurchasesDeferRequest}
\title{SubscriptionPurchasesDeferRequest Object}
\usage{
SubscriptionPurchasesDeferRequest(deferralInfo = NULL)
}
\arguments{
\item{deferralInfo}{The information about the new desired expiry time for the subscription}
}
\value{
SubscriptionPurchasesDeferRequest object
}
\description{
SubscriptionPurchasesDeferRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SubscriptionPurchasesDeferRequest functions: \code{\link{purchases.subscriptions.defer}}
}

