% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{SubscriptionPurchasesDeferResponse}
\alias{SubscriptionPurchasesDeferResponse}
\title{SubscriptionPurchasesDeferResponse Object}
\usage{
SubscriptionPurchasesDeferResponse(newExpiryTimeMillis = NULL)
}
\arguments{
\item{newExpiryTimeMillis}{The new expiry time for the subscription in milliseconds since the Epoch}
}
\value{
SubscriptionPurchasesDeferResponse object
}
\description{
SubscriptionPurchasesDeferResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
