% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Testers}
\alias{Testers}
\title{Testers Object}
\usage{
Testers(googleGroups = NULL, googlePlusCommunities = NULL)
}
\arguments{
\item{googleGroups}{No description}

\item{googlePlusCommunities}{No description}
}
\value{
Testers object
}
\description{
Testers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

