% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Timestamp}
\alias{Timestamp}
\title{Timestamp Object}
\usage{
Timestamp(nanos = NULL, seconds = NULL)
}
\arguments{
\item{nanos}{No description}

\item{seconds}{No description}
}
\value{
Timestamp object
}
\description{
Timestamp Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

