% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{Track}
\alias{Track}
\title{Track Object}
\usage{
Track(track = NULL, userFraction = NULL, versionCodes = NULL)
}
\arguments{
\item{track}{No description}

\item{userFraction}{No description}

\item{versionCodes}{No description}
}
\value{
Track object
}
\description{
Track Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
