% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{VoidedPurchase}
\alias{VoidedPurchase}
\title{VoidedPurchase Object}
\usage{
VoidedPurchase(purchaseTimeMillis = NULL, purchaseToken = NULL,
  voidedTimeMillis = NULL)
}
\arguments{
\item{purchaseTimeMillis}{The time at which the purchase was made, in milliseconds since the epoch (Jan 1, 1970)}

\item{purchaseToken}{The token that was generated when a purchase was made}

\item{voidedTimeMillis}{The time at which the purchase was cancelled/refunded/charged-back, in milliseconds since the epoch (Jan 1, 1970)}
}
\value{
VoidedPurchase object
}
\description{
VoidedPurchase Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A VoidedPurchase resource indicates a purchase that was either cancelled/refunded/charged-back.
}
