% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{VoidedPurchasesListResponse}
\alias{VoidedPurchasesListResponse}
\title{VoidedPurchasesListResponse Object}
\usage{
VoidedPurchasesListResponse(pageInfo = NULL, tokenPagination = NULL,
  voidedPurchases = NULL)
}
\arguments{
\item{pageInfo}{No description}

\item{tokenPagination}{No description}

\item{voidedPurchases}{No description}
}
\value{
VoidedPurchasesListResponse object
}
\description{
VoidedPurchasesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
