% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{UserComment}
\alias{UserComment}
\title{UserComment Object}
\usage{
UserComment(androidOsVersion = NULL, appVersionCode = NULL,
  appVersionName = NULL, device = NULL, deviceMetadata = NULL,
  lastModified = NULL, originalText = NULL, reviewerLanguage = NULL,
  starRating = NULL, text = NULL, thumbsDownCount = NULL,
  thumbsUpCount = NULL)
}
\arguments{
\item{androidOsVersion}{Integer Android SDK version of the user's device at the time the review was written, e}

\item{appVersionCode}{Integer version code of the app as installed at the time the review was written}

\item{appVersionName}{String version name of the app as installed at the time the review was written}

\item{device}{Codename for the reviewer's device, e}

\item{deviceMetadata}{Some information about the characteristics of the user's device}

\item{lastModified}{The last time at which this comment was updated}

\item{originalText}{Untranslated text of the review, in the case where the review has been translated}

\item{reviewerLanguage}{Language code for the reviewer}

\item{starRating}{The star rating associated with the review, from 1 to 5}

\item{text}{The content of the comment, i}

\item{thumbsDownCount}{Number of users who have given this review a thumbs down}

\item{thumbsUpCount}{Number of users who have given this review a thumbs up}
}
\value{
UserComment object
}
\description{
UserComment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
